requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
anthropic>=0.18.0
python-telegram-bot>=20.7
